[CHANGES] tinyInteger->integer

+++ database/migrations/2023_02_11_142353_create_departments_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('departments', function (Blueprint $table) {
             $table->id()->index();
             $table->string('department')->unique();
-            $table->tinyInteger('manager_user_fid')->nullable();
+            $table->integer('manager_user_fid')->nullable();
             $table->timestamps();
         });
     }
