+++ database/migrations/2023_02_19_132900_create_employees_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
     {
         Schema::create('employees', function (Blueprint $table) {
             $table->id()->index();
-            $table->tinyInteger('user_fid')->index();
-            $table->tinyInteger('contract_type_fid')->index();
-            $table->tinyInteger('designation_fid')->index();
-            $table->tinyInteger('department_fid')->nullable();
+            $table->integer('user_fid')->index();
+            $table->integer('contract_type_fid')->index();
+            $table->integer('designation_fid')->index();
+            $table->integer('department_fid')->nullable();
             $table->decimal('wage_per_year')->nullable();
             $table->timestamps();
         });
